+++ Entity/ContactForm.php
@@ -12,7 +12,7 @@ namespace c975L\ContactFormBundle\Entity;
 use Symfony\Component\Validator\Constraints as Assert;
 
 /**
- * Entity ContactForm (not linked to a db)
+ * Entity ContactForm (not linked to a DB)
  * @author Laurent Marquet <laurent.marquet@laposte.net>
  * @copyright 2017 975L <contact@975l.com>
  */